fix: 不要なカラムの削除
