+++ api/gateway/native/src/types/response/index.ts
@@ -1,4 +1,5 @@
 export * from './auth'
 export * from './book'
+export * from './bookshelf'
 export * from './error'
 export * from './user'
